pyproject.toml
[tool.poetry]
name = "rspace-client"
version = "2.6.1"
description = "A client for calling RSpace ELN and Inventory APIs"
license = "Apache-2.0"
authors = ["Research Innovations Ltd <support@researchspace.com>"]
maintainers = [
  "Matthias Kowalski <matthias@researchspace.com>",
  "Robert Lamacraft <robert.lamacraft@researchspace.com>"
]
readme = "README.md"
repository = "https://github.com/rspace-os/rspace-client-python"
documentation = "https://github.com/rspace-os/rspace-client-python"
keywords = ["RSpace", "ResearchSpace", "ELN", "Inventory"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Topic :: Software Development :: Libraries",
]
exclude = ["examples/*", "rspace_client/tests/*"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
fs = "^2.4.16"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"
black = "^21.6b0"
pytest = "^6.2.4"
playwright = "^1.54.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
